Adding GitTktFolder and GitTktCache classes
GitTktFolder is an interface of gitshelve with some code specific to ticketing.  It is responsible for managing the local numbering system within a specific folder.
GitTktCache is a shell for a caching system.  In the future, it will store all the loaded GitTktFolders and manage the cached database.  From here, any changes will be added to the appropriate folder and the cache database.  Any queries are done from the database unless the database is out of date.  In that case, it will be reloaded based on the data in the folders.