Fixing git user issue found on travis-ci.org
travis-ci was not working since git config user.name wouldn't work.
to set the unit tests to work in that environment AND a user's environment
without overridding the user's information, we will check if the
gitconfig exists.  If it does, we use it, if not, we write the user
name and email.